refactor: remove unused realm compatibility and tenant modules (#9)
- Remove unused compatibility.rs with legacy API methods
- Remove duplicate tenant module (741 lines)
- Update RealmConfig to use realm_rowid (i64) instead of realm_id (u32)
- Update Realm::save() to return i64 matching SQLite rowid type
- Update database schema: realmconfig.realm_id -> realmconfig.realm_rowid
- Fix ACL tests to use correct realm_id type
- Update supervit to use i64 for realm_rowid parameters

All tests passing, no breaking changes.

+++ actrix/crates/common/src/realm/model.rs
@@ -10,7 +10,7 @@ use sqlx::FromRow;
 ///
 #[derive(Debug, Clone, Serialize, Deserialize, Default, FromRow)]
 pub struct Realm {
-    pub rowid: Option<u32>,
+    pub rowid: Option<i64>,
 
     // 基础字段 - 所有服务都需要
     pub realm_id: u32,
